Extend IAttestationSigner: (#507)
- Add method to sign a attestation request"
