+++ src/WalletFramework.Oid4Vc/ClientAttestations/SignedWalletAttestationRequest.cs
@@ -1,0 +1,6 @@
+using Microsoft.IdentityModel.JsonWebTokens;
+using WalletFramework.Oid4Vc.WalletAttestations.Issuance;
+
+namespace WalletFramework.Oid4Vc.ClientAttestations;
+
+public record SignedWalletAttestationRequest(JsonWebToken SignedJwt, WalletAttestationRequest Value);
