+++ src/WalletFramework.Oid4Vc/WalletAttestations/Issuance/WalletAttestationResponse.cs
